Working map display in ascii
Note: Vehicle will be displayed if location entered is in the hallway.
If it is not in the hallway it will not return the vehicle location -
this is so that the vehicle is not reported to be inside a wall etc.

Script only needs to be fed an xy location to report on map, these
values correspond directly to the location values we recorded today.